pyproject.toml
[project]
name = "renamu-diccionario-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "chromadb>=1.5.7",
    "dashscope>=1.25.13",
    "dbfread>=2.0.7",
    "google-genai>=1.67.0",
    "google-generativeai>=0.8.6",
    "hdbscan>=0.8.42",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "openai>=2.28.0",
    "pandas>=3.0.1",
    "pdf2image>=1.17.0",
    "perustats>=0.1.7",
    "pillow>=12.1.1",
    "polars>=1.39.2",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "simpledbf>=0.2.6",
    "tables>=3.11.1",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
    "unidecode>=1.4.0",
]
